excel_parcer/requirements.txt
django
django-cors-headers
djangorestframework
python-decouple
celery
redis
pandas
sqlalchemy
django_redis
psycopg
psycopg2
psycopg-binary